Revise glossary for NUSContacts app
Updated the existing glossary, adding new terms
and updating definitions to enhance clarity and usefulness for
users. Includes the introduction of 'Tutor' as a combined term
for TAs and Profs, along with other key terms like
'Course Code', 'User Profile', and 'Search Function'. The aim is
to make the app more user-friendly and informative.